UI Revision, added Progressbar, added ToolTips
Revised the UI to be slimmer by using a Form layout. Texture Packer tab looks nice but the Texture Resizer tab has empty space that needs to be addressed in a later update. Integrated QProgressBar to inform user of current progress with operations. Also added ToolTips for various QWidgets to inform user.